feature: move single product view logic out from Products component into its own component. adds comments related to changes needed in Product model
