+++ client/react/components/Products.js
@@ -4,24 +4,10 @@ import { connect } from 'react-redux';
 import { Link } from 'react-router';
 
 import { selectProduct } from '../reducer/products';
-import localStore from 'store';
+
 // import Product from '/Product';
 
-const Products = ({ products, selectedProduct, setProduct }) => {
-  // should get products, selectedCategory??
-  console.log('ALL LOADED PRODUCTS', products);
-
-  function addToCart() {
-    const storeKey = String(selectedProduct.id);
-    const retrieved = localStore.get(storeKey);
-    if (!retrieved) {
-      localStore.set(storeKey, {quantity: 1, selectedProduct});
-    } else {
-      const newQuantity = retrieved.quantity + 1;
-      localStore.set(storeKey, {quantity: newQuantity, selectedProduct});
-    }
-  }
-
+const Products = ({ products, setProduct }) => {
   return (
     <div>
       {/* We can get category/search term from state?? */}
@@ -35,8 +21,6 @@ const Products = ({ products, selectedProduct, setProduct }) => {
           <button onClick={() => setProduct(product)}>Go To Product Page</button>
         </div>
       ))}
-
-      <button onClick={() => addToCart()}>Add Selected to Cart</button>
     </div>
   );
 };
@@ -50,7 +34,6 @@ const Products = ({ products, selectedProduct, setProduct }) => {
 const mapState = (state) => {
   return {
     products: state.products.allProducts,
-    selectedProduct: state.products.selectedProduct
   };
   //.filter(p => p.categories.includes(selectedCategory
 };
